Add `PyTreeHandler` for V1 API.
PiperOrigin-RevId: 734265601

+++ checkpoint/orbax/checkpoint/experimental/v1/_src/synchronization/types.py
@@ -21,5 +21,5 @@ T = TypeVar('T')
 
 class AsyncResponse(Protocol[T]):
 
-  def result(self, timeout: int | None = None) -> T:
+  def result(self, timeout: float | None = None) -> T:
     ...
